requirements.txt
torch
transformers
scikit-learn
matplotlib
seaborn
fastapi
uvicorn
pillow
torchvision